requirements.txt
asn1crypto==0.24.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
pycparser==2.19
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
whitenoise==4.1.3
pymongo==3.12.0
python-dotenv==0.19.0
pandas==1.3.5
dnspython==1.16.0